[Triton][Kernel] Refactor: organize triton activation kernels into layers/triton_ops
Create a folder layers/triton_ops for placing non-attention ops.

Move Triton kernel silu_and_mul_triton() to layers/triton_ops/activation.py
